+++ services/auth-server/cmd/users-list/users-list_test.go
@@ -33,7 +33,7 @@ func TestListRoute(t *testing.T) {
 		t.Fail()
 		return
 	}
-	defer cli.Drop()
+	defer cli.Close()
 
 	for _, test := range []struct {
 		name     string
@@ -71,7 +71,7 @@ func TestListRoute(t *testing.T) {
 		t.Run(test.name, func(u *testing.T) {
 			loadUsers(u, cli, test.ids)
 
-			r := server.GetRouter(TEST_DB, main.RouteBuilder)
+			r := server.GetRouter(cli.TargetConfig(), main.RouteBuilder)
 
 			req, _ := http.NewRequest("GET", "/", nil)
 			w := httptest.NewRecorder()
