+++ services/auth-server/cmd/users-post/users-post_test.go
@@ -37,7 +37,7 @@ func TestPostRoute(t *testing.T) {
 		t.Fail()
 		return
 	}
-	defer cli.Drop()
+	defer cli.Close()
 
 	for _, test := range []struct {
 		name     string
@@ -96,7 +96,7 @@ func TestPostRoute(t *testing.T) {
 		t.Run(test.name, func(u *testing.T) {
 			loadUsers(u, cli, test.exists)
 
-			r := server.GetRouter(TEST_DB, main.RouteBuilder)
+			r := server.GetRouter(cli.TargetConfig(), main.RouteBuilder)
 
 			data, _ := json.Marshal(test.args)
