--- services/auth-server/docker-compose.yaml
@@ -1,8 +1,0 @@
-version: '3.9'
-services:
-  cockroachdb:
-    image: cockroachdb/cockroach:v20.2.8
-    command: start-single-node --insecure
-    ports:
-    - "26257:26257"
-    - "8080:8080"
