+++ services/auth-server/internal/client/user.go
@@ -1,0 +1,10 @@
+package client
+
+// User - a struct representing a user
+type User struct {
+	ID       string
+	Email    string
+	Name     string
+	Password string
+	Verified bool
+}
