+++ services/go-auth-operator/crds/user.crd.yaml
@@ -1,0 +1,55 @@
+apiVersion: apiextensions.k8s.io/v1
+kind: CustomResourceDefinition
+metadata:
+  name: authusers.ponglehub.co.uk
+spec:
+  group: ponglehub.co.uk
+  subresources:
+    status: {}
+  versions:
+    - name: v1alpha1
+      served: true
+      storage: true
+      schema:
+        openAPIV3Schema:
+          type: object
+          properties:
+            spec:
+              type: object
+              properties:
+                name:
+                  type: string
+                email:
+                  type: string
+                password:
+                  type: string
+              required: [ name, email, password ]
+            status:
+              type: object
+              properties:
+                id:
+                  type: string
+                verified:
+                  type: boolean
+                  default: false
+      additionalPrinterColumns:
+      - name: Display
+        type: string
+        description: The user display name
+        jsonPath: .spec.name
+      - name: Email
+        type: string
+        description: The user email
+        jsonPath: .spec.email
+      - name: ID
+        type: string
+        description: The generated ID
+        jsonPath: .status.id
+      - name: Age
+        type: date
+        jsonPath: .metadata.creationTimestamp
+  scope: Namespaced
+  names:
+    plural: authusers
+    singular: authuser
+    kind: AuthUser
